requirements.txt
deltalake==0.20.0
duckdb==1.1.0
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
